# docs/swagger.yaml
definitions:
...
  controllers.OrderDetailInput:
    properties:
      jumlah:
        type: integer
      orderID:
        type: integer
      productID:
        type: integer
    type: object
...
  controllers.ProductInput:
    properties:
      ImageUrl:
        type: string
      categoryID:
        type: integer
      condition:
        type: string
      description:
        type: string
      heavy:
        type: string
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
    type: object
...
  models.OrderDetail:
    properties:
      createdAT:
        type: string
      harga:
        type: integer
      id:
        type: integer
      jumlah:
        type: integer
      orderId:
        type: integer
      productId:
        type: integer
      updatedAT:
        type: string
    type: object
  models.Product:
    properties:
      ImageUrl:
        type: string
      categoryID:
        type: integer
      condition:
        type: string
      createdAT:
        type: string
      description:
        type: string
      heavy:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
      updatedAT:
        type: string
      userID:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /orderDetails:
    post:
      description: Creating a new OrderDetail.
      parameters:
      - description: the body to create a new OrderDetail
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.OrderDetailInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrderDetail'
      security:
      - BearerToken: []
      summary: Create New OrderDetail.
      tags:
      - OrderDetail
  /orderDetails/{id}:
    delete:
      description: Delete a OrderDetail by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: OrderDetail id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one OrderDetail.
      tags:
      - OrderDetail
    patch:
      description: Update OrderDetail by id.
      parameters:
      - description: the body to update a new OrderDetail
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.OrderDetailInput'
      - description: OrderDetail id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OrderDetail'
      security:
      - BearerToken: []
      summary: Update OrderDetail.
      tags:
      - OrderDetail
...
  /products:
    get:
      description: Get list products refrence By userID
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      security:
      - BearerToken: []
      summary: Get Products By User Id
      tags:
      - Product
    post:
      description: Creating a new Product.
      parameters:
      - description: the body to create a new Product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.ProductInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      security:
      - BearerToken: []
      summary: Create New Product.
      tags:
      - Product
  /products/{id}:
    delete:
      description: Delete a Product by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Product.
      tags:
      - Product
    patch:
      description: Update Product by id.
      parameters:
      - description: the body to update a new Product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.ProductInput'
      - description: Product id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
      security:
      - BearerToken: []
      summary: Update Product.
      tags:
      - Product
...
swagger: "2.0"